requirements.txt
django==3.2.0 
djangorestframework==3.13.1
asgiref==3.5.2
django-filter==21.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
